+++ pytest/test_bytecode.py
@@ -129,8 +129,8 @@ def findlabels():
         assert offsets == [26, 25, 16]
 
     test_pyc = my_dir + "/../test/bytecode_2.7/01_dead_code.pyc"
-    (version, timestamp, magic_int, co, pypy, source_size, _) = load_module(test_pyc)
-    code = co.co_consts[0]
+    module = load_module(test_pyc)
+    code = module.co.co_consts[0]
     offsets = opcode_27.get_jump_targets(code, opcode_27)
     assert [10] == offsets
 
@@ -156,8 +156,8 @@ def findlabels():
     # Python 3.6 code wordcode
     # ------------------------
     test_pyc = my_dir + "/../test/bytecode_3.6/01_dead_code.pyc"
-    (version, timestamp, magic_int, co, pypy, source_size, _) = load_module(test_pyc)
-    code = co.co_consts[0]
+    module = load_module(test_pyc)
+    code = module.co.co_consts[0]
 
     #  2:           0 LOAD_FAST                 0 (a)
     #               2 POP_JUMP_IF_FALSE         8 (to 8)